rpmsg-lite: make timeout value to be aarch64 compatible
Some time related macro such as RL_BLOCK, RL_TRUE and RL_FALSE
are defined as "UL", so timeout value should be 64bit for aarch64
and 32bit for aarch32.

+++ lib/include/rpmsg_env.h
@@ -617,7 +617,7 @@ int32_t env_deinit_interrupt(void *env, int32_t vq_id);
  * @return RL_TRUE when link up, RL_FALSE when timeout.
  *
  */
-uint32_t env_wait_for_link_up(volatile uint32_t *link_state, uint32_t link_id, uint32_t timeout_ms);
+uint32_t env_wait_for_link_up(volatile uint32_t *link_state, uint32_t link_id, uintptr_t timeout_ms);
 
 /*!
  * env_tx_callback
